Rename property selector and controller
